refactor(react): use autogenerated utils
