+++ packages/react/src/components/react-component-lib/utils/attachProps.ts
@@ -24,11 +24,10 @@ export const attachProps = (node: HTMLElement, newProps: any, oldProps: any = {}
         const eventName = name.substring(2);
         const eventNameLc = eventName[0].toLowerCase() + eventName.substring(1);
 
-        if (typeof document !== 'undefined' && !isCoveredByReact(eventNameLc, document)) {
+        if (!isCoveredByReact(eventNameLc)) {
           syncEvent(node, eventNameLc, newProps[name]);
         }
       } else {
-        (node as any)[name] = newProps[name];
         const propType = typeof newProps[name];
         if (propType === 'string') {
           node.setAttribute(camelToDashCase(name), newProps[name]);
@@ -68,23 +67,27 @@ export const getClassName = (classList: DOMTokenList, newProps: any, oldProps: a
  * Checks if an event is supported in the current execution environment.
  * @license Modernizr 3.0.0pre (Custom Build) | MIT
  */
-export const isCoveredByReact = (eventNameSuffix: string, doc: Document) => {
-  const eventName = 'on' + eventNameSuffix;
-  let isSupported = eventName in doc;
+export const isCoveredByReact = (eventNameSuffix: string) => {
+  if (typeof document === 'undefined') {
+    return true;
+  } else {
+    const eventName = 'on' + eventNameSuffix;
+    let isSupported = eventName in document;
 
-  if (!isSupported) {
-    const element = doc.createElement('div');
-    element.setAttribute(eventName, 'return;');
-    isSupported = typeof (element as any)[eventName] === 'function';
+    if (!isSupported) {
+      const element = document.createElement('div');
+      element.setAttribute(eventName, 'return;');
+      isSupported = typeof (element as any)[eventName] === 'function';
+    }
+
+    return isSupported;
   }
-
-  return isSupported;
 };
 
 export const syncEvent = (
   node: Element & { __events?: { [key: string]: ((e: Event) => any) | undefined } },
   eventName: string,
-  newEventHandler?: (e: Event) => any,
+  newEventHandler?: (e: Event) => any
 ) => {
   const eventStore = node.__events || (node.__events = {});
   const oldEventHandler = eventStore[eventName];
@@ -101,7 +104,7 @@ export const syncEvent = (
       if (newEventHandler) {
         newEventHandler.call(this, e);
       }
-    }),
+    })
   );
 };
